feat: move components temp outside of ui dir
